+++ examples/ex/CustomGrid.tsx
@@ -11,7 +11,7 @@ const data = [
     {id: 3, product: 'Paper', type: '2', price: 5, stock: 2000},
     {id: 4, product: 'Chair', type: '1', price: 50, stock: 50},
     {id: 5, product: 'Computer', type: '1', price: 1000, stock: 20},
-    {id: 6, product: 'Rent', type: null, price: 2000, stock: undefined}
+    {id: 6, product: 'Rent', type: undefined, price: 2000, stock: undefined}
 ];
 
 // This is the schema
@@ -21,7 +21,7 @@ const schema: REL.Schema = [
     {
         name: 'type',
         type: [
-            {value: null, name: ''},
+            {value: undefined, name: ''},
             {value: '1', name: 'capex'},
             {value: '2', name: 'consumable'}
         ]