+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Support custom grid elements
 -   Support reference forwarding
 -   Support externally triggered data refresh
+-   Support custom fillers and default values
 
 # [1.0.0] 2021-05-02